Bugfixes from speedup

+++ Ensemble Learning/Bagging.py
@@ -30,7 +30,7 @@ class Bagging(object):
             res = pd.Series(res)
             res[res=='no']=-1.0
             res[res=='yes']=1.0
-            result = result.add(res)
+            result = result.add(res.astype(float))
         
         r.loc[result<0] = 'no'
         r.loc[result>=0] = 'yes'
